backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-dotenv==1.0.0
anthropic==0.7.7
pydantic==2.5.0
python-multipart==0.0.6 
pymongo==4.7.2
dnspython==2.4.2
pyautogen==0.2.0
reportlab==4.0.4
Pillow==10.0.1
jinja2==3.1.2
beautifulsoup4==4.12.2
websockets==12.0 